+++ integration/tezos/src/test/kotlin/com/rarible/protocol/union/integration/tezos/data/TezosDataFactory.kt
@@ -99,7 +99,7 @@ fun randomTezosOwnershipDto(itemId: ItemIdDto, creator: Part): TokenBalance {
     val ownershipId = itemId.toOwnership(creator.account)
     val (contract, tokenId) = CompositeItemIdParser.split(itemId.value)
     return TokenBalance(
-        id = randomInt(),
+        id = randomLong(),
         account = Alias(
             alias = null,
             address = ownershipId.owner.value
@@ -301,7 +301,7 @@ fun randomTzktToken(itemId: String) = Token(
 
 fun randomTzktTokenBalance(ownerId: OwnershipIdDto): TokenBalance {
     return TokenBalance(
-        id = randomInt(),
+        id = randomLong(),
         account = Alias(
             alias = null,
             address = ownerId.owner.value
